msg screen with logo
