+++ Vara-project/app/src/main/AndroidManifest.xml
@@ -9,10 +9,15 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
+        <activity android:name=".HomeMsg"></activity>
+        <activity android:name=".Home3" />
+        <activity android:name=".home_3" />
+
         <meta-data
             android:name="com.google.android.gms.ads.APPLICATION_ID"
-            android:value="@string/admob_app_id"/>
-        <activity android:name=".Home"></activity>
+            android:value="@string/admob_app_id" />
+
+        <activity android:name=".Home" />
         <activity android:name=".Register" />
         <activity android:name=".MainActivity">
             <intent-filter>
