+++ Vara-project/app/src/main/res/layout/activity_home.xml
@@ -4,10 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="#090808"
+    android:background="@drawable/logo3">
     tools:context=".Home"
     >
-
     <TextView
         android:id="@+id/dispTxt"
         android:layout_width="wrap_content"
