+++ Vara-project/app/src/main/res/layout/activity_home_msg.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="#090808"
+    tools:context=".HomeMsg">
+
+    <TextView
+        android:id="@+id/textView"
+        android:layout_width="467dp"
+        android:layout_height="112dp"
+        android:layout_marginStart="150dp"
+        android:layout_marginTop="425dp"
+        android:layout_marginEnd="183dp"
+        android:layout_marginBottom="743dp"
+        android:padding="5dp"
+        android:text="Welcome Message"
+        android:textAlignment="center"
+        android:textColor="#E72EBE"
+        android:textSize="36sp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
