+++ Vara-project/app/src/main/res/values/strings.xml
@@ -10,5 +10,5 @@
     <string name="Name" />
     <string name="signOut" />
     <string name="admob_app_id">ca-app-pub-9220432525424100~3672725629</string>
-    <string name="interstrial_add_id">ca-app-pub-9220432525424100/7017753851</string>
+    <string name="interstrial_add_id">ca-app-pub-3940256099942544/8691691433</string>
 </resources>
